requirements.txt
torch>=2.0.0
torchvision
transformers>=4.41.0
Pillow
opencv-python
numpy
scipy
matplotlib
tqdm
ipywidgets
notebook
jupyterlab
datasets
huggingface_hub
yt-dlp

einops
timm
